backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
sqlalchemy==2.0.25
aiosqlite==0.19.0
httpx==0.26.0
pillow
qrcode[pil]==7.4.2
reportlab==4.1.0
python-dotenv==1.0.0
numpy
tensorflow
jinja2==3.1.3
matplotlib
seaborn
scikit-learn
pandas
tensorflowjs
